[Feature] NLP trace support and example

+++ mqbench/fake_quantize/fixed.py
@@ -25,7 +25,7 @@ class FixedFakeQuantize(QuantizeBase):
         if self.fake_quant_enabled[0] == 1:
             if self.is_per_channel:
                 X = torch.fake_quantize_per_channel_affine(
-                    X, self.scale.data, self.zero_point.data.long(),
+                    X, self.scale.data, self.zero_point.data,
                     self.ch_axis, self.quant_min, self.quant_max)
             else:
                 X = torch.fake_quantize_per_tensor_affine(
